Splitting trpo_gym examples into two different examples using different policies.
For beginners using rllab it is unclear why an exception is thrown for the trpo_gym
example. It seems as though it should work as is. By splitting it up, we are help
guide beginners in understanding why different policies are needed and which ones
can be used for which environments.