Add ability to zip up an arbitrary amount of code folders when sending to ec2 instances.
This is the beginning of some changes so that we can package rllab as a library which
can be used externally. Now, you can have an arbitrary repository with your own
algorithms and can run something like:

run_experiment_lite(
    algo.train(),
    log_dir=None if args.use_ec2 else args.data_dir,
    # Number of parallel workers for sampling
    n_parallel=1,
    # Only keep the snapshot parameters for the last iteration
    snapshot_mode="last",
    # Specifies the seed for the experiment. If this is not provided, a random seed
    # will be used
    exp_prefix="myexperiments",
    seed=1,
    mode="ec2" if args.use_ec2 else "local",
    plot=False,
    # dry=True,
    added_project_directories=[osp.abspath(osp.join(osp.dirname(__file__), '.'))]
)

This will include your current project directory and send it to EC2 as if it's in the root dir of rllab.
Hence you can include your own libraries and run them without working in the rllab repo.